Separate explorer types into backend types and FE types
